All files are seperated and ready to be implemented
