+++ src/client.cpp
@@ -2,5 +2,4 @@
 
 #include <string>
 
-
 Client::Client(std::string name ) : clientName(name) {}
