+++ src/console_ui.cpp
@@ -1,6 +1,4 @@
 #include "console_ui.h"
 
-#include <stdlib.h>
-#include <iostream>
-#include <string>
 
+// To be implemented
